// frontend/src/utils/models/location.d.ts
export type TClientLocation = {
  lat: string;
  lon: string;
  timeZone: string;
  city: string;
  country: string;
};

export type TServiceLocation = {
  lat: string;
  lon: string;
  timeZone: string;
  city: string;
  country: string;
  region: string;
  regionName: string;
  status: string;
  countryCode: string;
  zip: string;
};